% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{as_tibbler}
\alias{as_tibbler}
\title{func_common}
\usage{
as_tibbler(x, coln = NULL)
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that could reasonably be coerced to a tibble.}

\item{coln}{A strings vector specifying column name.}
}
\description{
R function file for 'andsr' (Analysis of Nonlinear Dynamical Systems in R).
This file contains common functions throuhout \pkg{andsr}.
Initially written on 20210319 by K.Kawatsu.
Last update: 20230120.
Modified version of as_tibble
}
