% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{find_best_dim}
\alias{find_best_dim}
\title{Grid search of best embedding dimension with simplex projection}
\usage{
find_best_dim(
  ts,
  cols,
  lib = matrix(c(1, nrow(ts)), nrow = 1),
  pred = NULL,
  range = 1:10,
  Tps = rep(1, length(cols)),
  criterion = "rmse",
  both = TRUE,
  threadNo = detectCores()
)
}
\arguments{
\item{ts}{A vector, matrix, data.frame or tibble, which contains time-series data.}

\item{cols}{A vector, which sets the columns to be analyzed.}

\item{lib}{Vector or 2-column matrix, which sets the knot positions in library data.}

\item{pred}{Vector or 2-column matrix, which sets the knot position in test data.}

\item{range}{Integer vector, which sets the range of dimension search.}

\item{Tps}{Integer vector, which setse the prediction time horison for each variable.}

\item{criterion}{Strings, which switches the cost function (Rho, MAE or RMSE)
for the dimension search (default = "rmse").}

\item{both}{Logical, determine whether backward prediction is performed.}

\item{threadNo}{Integer, which sets the thread number of parallel computing.}
}
\description{
\code{find_best_dim} returns the best embedding dimension.
}
