% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{find_knot}
\alias{find_knot}
\title{func_ntsa}
\usage{
find_knot(ts, key_col = NULL, time_col = 1, diff = 1)
}
\arguments{
\item{ts}{data.frame or tibble contains the time-series data.}

\item{key_col}{Numeric or strings, which contains the labels of the data group.}

\item{time_col}{Numeric or strings, which specifies the time-point column.}

\item{diff}{Numeric, which sets the value of increment step in 'time_col'.}
}
\description{
R function file for 'andsr' (Analysis of Nonlinear Dynamical Systems in R).
This file contains functions related to nonlinear time-series analysis.
Initially written on 20210706 by K.Kawatsu.
Last update: 20240531.
Find knot positions in data

\code{find_knot} returns a 2-column matrix, where start and end point of an i'th knot in ts.
}
