% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{gen_emat}
\alias{gen_emat}
\title{Embedding matrinx generation with multivariate and multi-timelags}
\usage{
gen_emat(ts, cols, lags, knot = matrix(c(1, nrow(ts)), nrow = 1))
}
\arguments{
\item{ts}{vector, matrix, data.frame or tibble
which contains time series to make an embedding matrix.}

\item{cols}{Numeric or charactor vector,
which selects the ts's column for the embedding matrix reconstruction.}

\item{lags}{Integer vector, which sets
the time delay value for each coordinate in embedding matrix.}

\item{knot}{vector or 2-column matrix, which specifies the knot position in ts.}
}
\description{
\code{gen_emat} return an embedding matrix.
}
