% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{gen_map}
\alias{gen_map}
\title{Calculate discrete-time time series data}
\usage{
gen_map(x, func, params, len, p_noise = 0, o_noise = 0)
}
\arguments{
\item{x}{Numeric vector ,set initial states.}

\item{func}{Strings, which specifies the specific equation to be calculated.}

\item{params}{List, containing required parameters in 'func'.}

\item{len}{Integer, the time-series length to be generated.}

\item{p_noise}{Numeric, process-noise strength.}

\item{o_noise}{Numeric, observation-error strength.}
}
\description{
Calculate discrete-time time series data
}
