% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_surrogate.R
\name{gen_surr_rs}
\alias{gen_surr_rs}
\title{Surrogate data generation with random shuffling}
\usage{
gen_surr_rs(
  X,
  col,
  knot = matrix(c(1, nrow(X)), nrow = 1),
  iter = 100,
  seed = NULL
)
}
\arguments{
\item{X}{data matrix.}

\item{col}{Integer, the column position of surrogaate target variable in 'X'.}

\item{knot}{Vector or 2-column matrix, which specifies the knot positions in the data.}

\item{iter}{Integer, No. of surrogate generation.}

\item{seed}{Integer, RNG seed.}
}
\description{
\code{gen_surr_rs} returns a column-wise piled surrogate data with RS algorithm.
}
