% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_cv}
\alias{get_cv}
\title{Calculation of CV (Coefficient of Variance)}
\usage{
get_cv(x)
}
\arguments{
\item{x}{Numeric, input vector.}
}
\description{
\code{get_cv} returns the CV of input vector x.
}
