% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_lmd}
\alias{get_lmd}
\title{R wrapper function of getLEDist in ands_common.cpp}
\usage{
get_lmd(X, theta)
}
\arguments{
\item{X}{Data matrix.}

\item{theta}{A matrix, in which elements capture the nonlinearity of each column in X.}
}
\description{
\code{get_lmd} returns the LMD (Local Manifold Distance) of the matrix X.
}
