% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_mae}
\alias{get_mae}
\title{R wrapper function of getMAE in ands_common.cpp}
\usage{
get_mae(x, y)
}
\arguments{
\item{x}{Numeric vector 1.}

\item{y}{Numeric vector 2.}
}
\description{
\code{get_mae} returns the MAE (Mean Absolute Error) between x and y.
}
