% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{get_mvd}
\alias{get_mvd}
\title{Calculation of Multiview-Distance (MVD) matrix}
\usage{
get_mvd(
  ts,
  cols,
  knot = matrix(c(1, nrow(ts)), nrow = 1),
  E,
  Tp,
  lmax,
  threadNo = detectCores()
)
}
\arguments{
\item{ts}{vector, matrix, data.frame or tibble
which contains time series to make an embedding matrix.}

\item{cols}{Numeric or charactor vector,
which selects the ts's column for the embedding matrix reconstruction.}

\item{knot}{vector or 2-column matrix, which specifies the knot position in ts.}

\item{E}{Integer, which sets the embedding dimension.}

\item{Tp}{Integer, which sets the prediction-time horizon.}

\item{lmax}{Integer, which sets the maximum value of timelags.}

\item{threadNo}{Integer, which is the No. of cores used for parallel computing.}
}
\description{
\code{get_mvd} returns MVD matrix calculated with an ensemble of euclidean
distance matrix of the top-k embeeding in multivariate simplex projection.
Algorithm is adopted from Chang et al. (2021), Ecol Lett.
For more details, see also \code{\link[rEDM]{multiview}}.
}
