% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_pred}
\alias{get_pred}
\title{R wrapper function of getPred in ands_common.cpp}
\usage{
get_pred(C, X)
}
\arguments{
\item{C}{Model coefficient matrix.}

\item{X}{Data matrix.}
}
\description{
\code{get_pred} returns a prediction vector calculated with Data matrix and model coefficient matrix.
}
