% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_rho}
\alias{get_rho}
\title{R wrapper function of getRho in ands_common.cpp}
\usage{
get_rho(x, y)
}
\arguments{
\item{x}{Numeric vector 1.}

\item{y}{Numeric vector 2.}
}
\description{
\code{get_rho} returns the correlation coefficient between x and y.
}
