% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_rmse}
\alias{get_rmse}
\title{func_common}
\usage{
get_rmse(x, y)
}
\arguments{
\item{x}{Numeric vector 1.}

\item{y}{Numeric vector 2.}
}
\description{
R function file for 'andsr' (Analysis of Nonlinear Dynamical Systems in R).
This file contains common functions throuhout \pkg{andsr}.
Initially written on 20210319 by K.Kawatsu.
Last update: 20220428.
R wrapper function of getRMSE in ands_common.cpp

\code{get_rmse} returns the RMSE (Root Mean Squared Error) between x and y.
}
