% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{get_trunc}
\alias{get_trunc}
\title{r wrapper function of getoptht in ands_ntsa.cpp}
\usage{
get_trunc(sv, m, n)
}
\arguments{
\item{sv}{Numeric vectors for singular values.}

\item{m}{Integer, size of measurements.}

\item{n}{Integer, data size}
}
\description{
\code{get_trunc} returns the truncated rank of SVD with optimal hard threshold.
}
