% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{LMDmap}
\alias{LMDmap}
\title{R wrapper function of lmdSMap in ands_ntsa.cpp}
\usage{
LMDmap(
  X,
  col = 1,
  lib = matrix(c(1, nrow(X)), nrow = 1),
  pred = NULL,
  Tp = 1,
  method,
  theta,
  lambda
)
}
\arguments{
\item{X}{An embedding matrix, which is constructed by the function \code{\link{gen_emat}}.}

\item{col}{Integer, which specifides the position of prediction target in ts.}

\item{lib}{Vector or 2-column matrix, which sets the knot positions in library data.}

\item{pred}{Vector or 2-column matrix, which sets the knot position in test data.}

\item{Tp}{Integer, which is the prediction-time horizon.}

\item{method}{Integer, which switches the LMD-based mapping solver.
If col == 0, y = C X is solved with SVD, otherwise is solved with L2 norm penelization.}

\item{theta}{A matrix, where each element controls the locality of coordinates in X.}

\item{lambda}{Numeric, which controls the penalization strength in Ridge regression.}
}
\description{
\code{LMDmap} returns Jacobian elements estimated by LMD-based nonlinear regression.
}
