% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{LMDsmplx}
\alias{LMDsmplx}
\title{R wrapper function of lmdSmplx in ands_ntsa.cpp}
\usage{
LMDsmplx(
  ts,
  col,
  lib = matrix(c(1, nrow(ts)), nrow = 1),
  pred = NULL,
  E,
  Tp = 1,
  theta = NULL,
  both = TRUE
)
}
\arguments{
\item{ts}{A vector, matrix, data.frame or tibble, which contains time-series data.}

\item{col}{Integer, which specifides the position of prediction target in ts.}

\item{lib}{Vector or 2-column matrix, which sets the knot positions in library data.}

\item{pred}{Vector or 2-column matrix, which sets the knot position in test data.}

\item{E}{Integer, embedding dimension.}

\item{Tp}{Integer, which is the prediction-time horizon.}

\item{theta}{A matrix, where each element controls the locality of coordinates in ts.}

\item{both}{Logical, determine whether backward prediction is performed.}
}
\description{
\code{LMDsmplx} returns stats and predictions by LMD-based simplex projection.
}
