% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{logistic}
\alias{logistic}
\title{Coupled logistic model}
\usage{
logistic(state, r, A)
}
\arguments{
\item{state}{Numeric vector (n-dimensional), containing temporal state of N-species density.}

\item{r}{Numeric vector (n-dimensional), containing intrinsic growth rate.}

\item{A}{NUmeric matrix (n-n dimensional), containing interaction strength between species i and j.}
}
\description{
Coupled logistic model
}
