% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{lorenz}
\alias{lorenz}
\title{Lorenz63 system}
\usage{
lorenz(state, sigma, beta, rho)
}
\arguments{
\item{state}{Numeric vector (3-dimensional), containing temporal state of x, y and z in Lorenz63 system.}

\item{sigma}{Numeric, corresponding to the parameter sigma in Lorenz63 system.}

\item{beta}{Numeric, corresponding to the paramter beta in Lorenz63 system.}

\item{rho}{Numeric, corresponding to the parameter rho in Lorenz63 system.}
}
\description{
Lorenz63 system
}
