% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{multiview}
\alias{multiview}
\title{Wrapper of Multiview in rEDM}
\usage{
multiview(
  ts,
  cols,
  tar = 1,
  lib = matrix(c(1, nrow(ts)), nrow = 1),
  pred = NULL,
  E,
  Tp = 1,
  lmax,
  threadNo = detectCores()
)
}
\arguments{
\item{ts}{vector, matrix, data.frame or tibble
which contains time series to make an embedding matrix.}

\item{cols}{Numeric or character vector,
which selects the ts's column for the embedding matrix reconstruction.}

\item{tar}{Integer or Strings, which sets the columns to be predicted.}

\item{lib}{Vector or 2-column matrix, which sets the knot positions in library data.}

\item{pred}{Vector or 2-column matrix, which sets the knot positions in predition data.}

\item{E}{Integer, which sets the embedding dimension.}

\item{Tp}{Integer, which sets the prediction-time horizon.}

\item{lmax}{Integer, which sets the maximum value of timelags.}

\item{threadNo}{Integer, which is the No. of cores used for parallel computing.}
}
\description{
\code{multiview} returns the results of multiview-embedding.
For more details, see \code{\link[rEDM]{Multiview}}
}
