% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{odeRK4}
\alias{odeRK4}
\title{ODE solver with 4-th order Runge-Kutta approximation}
\usage{
odeRK4(x, func, params, len, h = 0.01)
}
\arguments{
\item{x}{Numeric vector ,set initial states.}

\item{func}{Strings, which specifies the specific equation to be calculated.}

\item{params}{List, containing required parameters in 'func'.}

\item{len}{Integer, the time-series length to be generated.}

\item{h}{Numeric, time-step paramter in RK-approximation}
}
\description{
ODE solver with 4-th order Runge-Kutta approximation
}
