% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{pullist}
\alias{pullist}
\title{Accessor function of list data}
\usage{
pullist(lst, name)
}
\arguments{
\item{lst}{A list to be accessed.}

\item{name}{A strings specifying the data to be accessed.}
}
\description{
Accessor function of list data
}
