% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{set_cnames}
\alias{set_cnames}
\title{Function to set column names within pipes}
\usage{
set_cnames(x, nm)
}
\arguments{
\item{x}{tibble, data.frame or matrix to be renamed}

\item{nm}{Strings vector for the column name}
}
\description{
Function to set column names within pipes
}
