% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{shift}
\alias{shift}
\title{Shift vector elements forward or backward depending on the lag value}
\usage{
shift(x, lag)
}
\arguments{
\item{x}{An input vector.}

\item{lag}{Integer to set the lag value for bwd/fwd shifts.}
}
\description{
\code{shift} shifts an input vector x bacward or forward.
}
