% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{smap}
\alias{smap}
\title{R wrapper function of smapSVD in ands_ntsa.cpp}
\usage{
smap(X, y, idx_l, idx_p, coef, D, theta)
}
\arguments{
\item{X}{A matrix, which is constructed by the function \code{\link{gen_emat}}.}

\item{y}{A vector, which is constructed by the function \code{\link{gen_emat}}.}

\item{idx_l}{A vector specifying valid positions in library IDs.}

\item{idx_p}{A vector specifying valid positions in prediction IDs.}

\item{coef}{A matrix storing model coefficients.}

\item{D}{A distant matrix.}

\item{theta}{A uniform localization parameter.}
}
\description{
R wrapper function of smapSVD in ands_ntsa.cpp
}
