% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{smap_geodes}
\alias{smap_geodes}
\title{S-map with geodesic distance}
\usage{
smap_geodes(
  X,
  col = 1,
  lib = matrix(c(1, nrow(X)), nrow = 1),
  pred = NULL,
  Tp = 1,
  dmat = NULL,
  seed = NULL,
  threadNo = detectCores(),
  iter = 100,
  criterion = "rmse",
  k_range = 1:10,
  inits = c(1, 1),
  sigmas = c(0.05, 0.01),
  drop_dist = TRUE
)
}
\arguments{
\item{X}{An embedding matrix.}

\item{col}{An integer specifying the position of prediction target in X.}

\item{lib}{A vector or 2-column matrix, which sets the knot positions in library data.}

\item{pred}{A vector or 2-column matrix, which sets the knot positions in predition data.}

\item{Tp}{An integer, the prediction-time horizon.}

\item{dmat}{A distance matrix (If NULL, Normal-Euclidean distance is used for the analysis).}

\item{seed}{An integer for the RNG seed.}

\item{threadNo}{An integer specifying the number of threads of parallel computing.}

\item{iter}{An integer specifying the iteration number.}

\item{criterion}{A strings switching the criterion for the best model evaluation.}

\item{inits}{A double vector specifying the initial values of delta and theta.}

\item{sigmas}{A double vector specifyign the mutation rate of delta and theta.}

\item{krange}{An integer vector specifying number of neighbor points in reconstructing geodesic distance.}
}
\description{
\code{smap_geodes} returns the results of S-map with geodesic distance
Implementation is based on the algorithm in Qu et al. 2023, Ecol. Indic.
}
