% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{smap_mulvar}
\alias{smap_mulvar}
\title{S-map with multivariate embedding (a.k.a. multivariate S-map)}
\usage{
smap_mulvar(
  X,
  col = 1,
  lib = matrix(c(1, nrow(X)), nrow = 1),
  pred = NULL,
  Tp = 1,
  dmat = NULL,
  range = seq(0, 10, 1),
  threadNo = detectCores(),
  criterion = "rmse"
)
}
\arguments{
\item{X}{A matrix, which is constructed by the function \code{\link{gen_emat}}.}

\item{col}{Integer, which specifies the position of prediction target in X.}

\item{lib}{Vector or 2-column matrix, which sets the knot positions in library data.}

\item{pred}{Vector or 2-column matrix, which sets the knot position in test data.}

\item{Tp}{Integer, which is the prediction-time horizon.}

\item{dmat}{Distance matrix to calculate local weights.
If dmat is obtained by \code{\link{get_mvd}}, the procedure follows
the MDR (MultiView-Distance Regularized) S-map algorithm (see Chang et al. 2021, Ecol Lett.)
If NULL< then Normal-Eculidean Distance is instead used for the analysis.}

\item{range}{Integer vector, which sets the range of dimension search.}

\item{threadNo}{Integer, which sets the thread number of parallel computing.}

\item{criterion}{Strings, which switches the cost function (Rho, MAE or RMSE)
for the dimension search (default = "rmse").}
}
\description{
\code{smap_mulvar} returns the (tibble) result of multivariate S-map.
Algorithm is adopted from Deyle et al. (2016) PRSB.
}
