% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ntsa.R
\name{smplx}
\alias{smplx}
\title{R wrapper function of lmdSmplx in ands_ntsa.cpp version 2}
\usage{
smplx(
  X,
  col = 1,
  lib = matrix(c(1, nrow(X)), nrow = 1),
  pred = NULL,
  Tp = 1,
  dmat = NULL
)
}
\arguments{
\item{X}{A matrix, which is constructed by the function \code{\link{gen_emat}}.}

\item{col}{Integer, which specifides the position of prediction target in ts.}

\item{lib}{Vector or 2-column matrix, which sets the knot positions in library data.}

\item{pred}{Vector or 2-column matrix, which sets the knot position in test data.}

\item{Tp}{Integer, which is the prediction-time horizon.}

\item{dmat}{A distant matrix to identify neighorhood relationshiop.}
}
\description{
R wrapper function of lmdSmplx in ands_ntsa.cpp version 2
}
