% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andsr_package.R
\docType{package}
\name{andsr}
\alias{andsr}
\title{Applications of analysis of nonlinear dynamical systems in R.}
\description{
The \pkg{andsr} package is a new implementation of andsr methods,
including the functions of nonlinear time-series analysis (ands_ntsa),
those of lyapunov exponent/vector analysis (ands_lyap) and
those of controllability analysis (ands_ctrb).
The \pkg{andsr} will be applicable to diverse purposes, including
nonlinear time-series forecasting, system identification, stability and keystone analysis.
}
