% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{gen_map}
\alias{gen_map}
\title{Calculate discrete-time time series data}
\usage{
gen_map(x_ini, fun, params, input = NULL, t_end, o_noise = 0)
}
\arguments{
\item{x_ini}{A numeric vector, set the initial states.}

\item{fun}{A functional object specifying the type of differential equation.}

\item{params}{A list containing required parameters in 'fun'.}

\item{t_end}{A double specifying the end time at the termination of calculation.}

\item{o_noise}{A numeric, observation error strength/}
}
\description{
Calculate discrete-time time series data
}
