% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_common.R
\name{odeRK4}
\alias{odeRK4}
\title{ODE solver with 4-th order Runge-Kutta approximation}
\usage{
odeRK4(x_ini, fun, params, input = NULL, t_end, h = 0.01)
}
\arguments{
\item{x_ini}{A numeric vector, set the initial states.}

\item{fun}{A functional object specifying the type of differential equation.}

\item{params}{A list containing required parameters in 'fun'.}

\item{t_end}{A double specifying the end time at the termination of calculation.}

\item{h}{A double, specifying integration time in RK-approximation.}
}
\description{
ODE solver with 4-th order Runge-Kutta approximation
}
